refactor: helper function to generate logs
