+++ e2e/packages/test-data/generate-test-data-big.ts
@@ -2,23 +2,9 @@ import fs from "node:fs/promises";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { startProxy } from "@viem/anvil";
-import { execa } from "execa";
-import {
-  ClientConfig,
-  createPublicClient,
-  createWalletClient,
-  encodeEventTopics,
-  getContract,
-  http,
-  isHex,
-  numberToHex,
-} from "viem";
-import { mudFoundry } from "@latticexyz/common/chains";
-import { storeEventsAbi } from "@latticexyz/store";
-import { privateKeyToAccount } from "viem/accounts";
-import IWorldAbi from "../contracts/out/IWorld.sol/IWorld.abi.json";
+import { generateLogs } from "./generateLogs";
 
-const SIZES = [10, 100, 1000];
+const COUNTS = [10, 100, 1000];
 
 console.log("starting proxy");
 await startProxy({
@@ -29,86 +15,19 @@ await startProxy({
   },
 });
 
-for (let i = 0; i < SIZES.length; i++) {
-  const size = SIZES[i];
+for (let i = 0; i < COUNTS.length; i++) {
+  const count = COUNTS[i];
+
+  // each group of logs gets a fresh anvil instance
+  const rpc = `http://127.0.0.1:8545/${count}`;
+
+  console.log(`generating logs for ${count} records`);
+  const logs = await generateLogs(count, rpc);
 
   const logsFilename = path.join(
     path.dirname(fileURLToPath(import.meta.url)),
-    `../../../test-data/world-logs-${size}.json`
+    `../../../test-data/world-logs-${count}.json`
   );
-
-  // each set of logs gets a fresh anvil instance
-  const rpc = `http://127.0.0.1:8545/${size}`;
-
-  console.log("deploying world");
-  const { stdout, stderr } = await execa("pnpm", ["mud", "deploy", "--rpc", rpc, "--saveDeployment", "false"], {
-    cwd: "../contracts",
-    stdio: "pipe",
-    env: {
-      DEBUG: "mud:*",
-    },
-  });
-  if (stderr) console.error(stderr);
-  if (stdout) console.log(stdout);
-
-  const [, worldAddress] = stdout.match(/worldAddress: '(0x[0-9a-f]+)'/i) ?? [];
-  if (!isHex(worldAddress)) {
-    throw new Error("world address not found in output, did the deploy fail?");
-  }
-  console.log("got world address", worldAddress);
-
-  const clientOptions = {
-    chain: mudFoundry,
-    transport: http(rpc),
-    pollingInterval: 1000,
-  } as const satisfies ClientConfig;
-
-  const publicClient = createPublicClient(clientOptions);
-
-  // anvil default private key
-  const account = privateKeyToAccount("0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80");
-  const walletClient = createWalletClient({
-    ...clientOptions,
-    account,
-  });
-
-  // use viem Contract instance to avoid nonce management
-  const worldContract = getContract({
-    address: worldAddress,
-    abi: IWorldAbi,
-    publicClient,
-    walletClient,
-  });
-
-  console.log(`Setting ${size} records`);
-  for (let i = 0; i < size - 1; i++) {
-    await worldContract.write.setNumber([i, i]);
-  }
-
-  const lastTx = await worldContract.write.setNumber([size, size]);
-
-  console.log("waiting for tx");
-  const receipt = await publicClient.waitForTransactionReceipt({ hash: lastTx });
-
-  console.log("fetching logs", receipt.blockNumber);
-  const logs = await publicClient.request({
-    method: "eth_getLogs",
-    params: [
-      {
-        address: worldAddress,
-        topics: [
-          storeEventsAbi.flatMap((event) =>
-            encodeEventTopics({
-              abi: [event],
-              eventName: event.name,
-            })
-          ),
-        ],
-        fromBlock: numberToHex(0n),
-        toBlock: numberToHex(receipt.blockNumber),
-      },
-    ],
-  });
 
   console.log("writing", logs.length, "logs to", logsFilename);
   await fs.writeFile(logsFilename, JSON.stringify(logs, null, 2));
